requirements.txt
axicli @ https://cdn.evilmadscientist.com/dl/ad/public/AxiDraw_API.zip#sha256=c29ef0792dc8a2006a3a4abcb306e8d7fa5b93f8ce83c0c781eed226d7eeca24
axidrawinternal @ file:///C:/Users/harsh/AppData/Local/Temp/pip-req-build-_zc3jn4k/prebuilt_dependencies/axidrawinternal-3.9.6-py2.py3-none-any.whl#sha256=744be32fa64d85adb9dee4d03098af272be4bbf04a0c3dee40b685c3b0d97c4e
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
idna==3.11
ImageIO==2.37.2
ink_extensions==2.2.0
kiwisolver==1.4.9
lazy_loader==0.4
lxml==6.0.2
matplotlib==3.10.8
networkx==3.6.1
noise==1.2.2
numpy==2.4.0
packaging==25.0
pillow==12.0.0
plotink==1.14.0
pyclipper==1.4.0
pyparsing==3.3.1
pyserial==3.5
python-dateutil==2.9.0.post0
requests==2.32.5
scikit-image==0.26.0
scipy==1.16.3
shapely==2.1.2
six==1.17.0
svgwrite==1.4.3
tifffile==2025.12.20
tqdm==4.67.1
urllib3==2.6.2
